marketing/requirements.txt
pandas
pytrends
tweepy
newsapi-python
matplotlib
Pillow